Update node creation to use SilverStripe naming conventions Get it to actually work
