+++ .gitignore
@@ -1,5 +1,3 @@
 /__tests__
 /yarn.lock
-/*.js
-!index.js
 node_modules/
