+++ plugin-options.js
