--- src/index.js
@@ -1,3 +1,0 @@
-
-console.log('plugins folder doesnt work');
-module.exports.schemes = require('./src/schemes.js');
